+++ CHANGELOG.md
@@ -143,6 +143,7 @@
   [3519]
 - [Removed obsolete `from_xls` and `from_xlsx` functions. Added support for
   reading column names from first row in `File_Format.Excel`][3523]
+- [Added `File_Format.Delimited` support to `Table.write` for new files.][3528]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -225,6 +226,7 @@
 [3518]: https://github.com/enso-org/enso/pull/3518
 [3519]: https://github.com/enso-org/enso/pull/3519
 [3523]: https://github.com/enso-org/enso/pull/3523
+[3528]: https://github.com/enso-org/enso/pull/3528
 
 #### Enso Compiler